requirements/base.txt
Django<2.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django-extensions==2.0.7
django-ratelimit==1.1.0
django-smartfields==1.0.9
mysqlclient
Pillow==5.2.0
psycopg2
pytz==2018.5
requests
six==1.11.0
sorl-thumbnail==12.4.1
